+++ node/src/main/kotlin/net/corda/node/Corda.kt
@@ -110,7 +110,7 @@ fun main(args: Array<String>) {
     log.info("VM ${info.vmName} ${info.vmVendor} ${info.vmVersion}")
     log.info("Machine: ${InetAddress.getLocalHost().hostName}")
     log.info("Working Directory: ${cmdlineOptions.baseDirectory}")
-    log.info("Starting as node on ${conf.artemisAddress}")
+    log.info("Starting as node on ${conf.messagingAddress}")
 
     try {
         cmdlineOptions.baseDirectory.createDirectories()