+++ test-utils/src/main/kotlin/net/corda/testing/node/NodeBasedTest.kt
@@ -73,7 +73,7 @@ abstract class NodeBasedTest {
                 rpcUsers,
                 mapOf(
                         "networkMapService" to mapOf(
-                                "address" to networkMapNode.configuration.artemisAddress.toString(),
+                                "address" to networkMapNode.configuration.messagingAddress.toString(),
                                 "legalName" to networkMapNode.info.legalIdentity.name
                         )
                 ) + configOverrides
@@ -121,7 +121,7 @@ abstract class NodeBasedTest {
                 allowMissingConfig = true,
                 configOverrides = mapOf(
                         "myLegalName" to legalName,
-                        "artemisAddress" to freeLocalHostAndPort().toString(),
+                        "messagingAddress" to freeLocalHostAndPort().toString(),
                         "rpcAddress" to freeLocalHostAndPort().toString(),
                         "extraAdvertisedServiceIds" to advertisedServices.map { it.toString() },
                         "rpcUsers" to rpcUsers.map {